Изменил способ отображения стори, добавил SideBySide отображение тестового компонента
