--- src/components/TestButton/TestButton.stories.js
@@ -1,21 +1,0 @@
-import TestButton from "./TestButton";
-
-export default { component: TestButton};
-
-
-
-export const Default  = {}
-
-export const LightTheme = {
-    ...Default
-}
-LightTheme.parameters = {
-    theme: 'light',
-}
-
-export const DarkTheme = {
-    ...Default
-}
-DarkTheme.parameters = {
-    theme: 'dark',
-}
